Extract Balances component
